octofit-tracker/backend/requirements.txt
Django==4.1
djongo==1.3.6
djangorestframework==3.14.0
pymongo==3.12
django-cors-headers==4.5.0
